avoid conflicts among use-statements
