+++ libs/gesha-core/src/targets/rust_type/media_type_def.rs
@@ -6,6 +6,3 @@ pub struct MediaTypeDef {
     /// e.g. "ApplicationJson" -> "application/json"
     pub translator: IndexMap<EnumVariantName, &'static str>,
 }
-
-// TODO: use EnumVariantName instead of &str
-// TODO: define MediaTypeValue
